rusty_ache::engine::config

Trait Config

pub trait Config {
    // Required methods
    fn set_resolution(&mut self, width: u32, height: u32);
    fn get_resolution(&self) -> Resolution;
    fn new(resolution: Resolution) -> Self
       where Self: Sized;
}

Trait defining configuration interface for engine settings.

Allows setting and retrieving the screen resolution. The new method constructs a configuration instance with an initial resolution.

Required Methods

fn set_resolution(&mut self, width: u32, height: u32)

Sets the resolution width and height.

Parameters

width: Screen width in pixels.

height: Screen height in pixels.

fn get_resolution(&self) -> Resolution

Gets the current resolution.

Returns

A Resolution struct representing the current screen dimensions.

fn new(resolution: Resolution) -> Self
where Self: Sized,

Creates a new configuration instance with the given resolution.

Parameters

resolution: Initial resolution settings.

Returns

A new implementing instance of the Config trait.