rusty_ache::engine::scene::game_object::components::script

Trait Script

pub trait Script {
    // Required methods
    fn action(&mut self, game_object: &mut GameObject);
    fn new(is_downed: bool) -> Self
       where Self: Sized;
    fn clone_box(&self) -> Box<dyn Script + Send + Sync>;
}

Trait representing a script that can be attached to a game object.

Scripts encapsulate game logic that modifies the game object’s state. This trait requires the implementation of:

an action method performing operations on a mutable reference to GameObject

a new constructor with an initialization parameter is_downed

Implementors can define custom behavior for initialization and per-frame updates.

Required Methods

fn action(&mut self, game_object: &mut GameObject)

Apply the script’s action on the given mutable game object reference.

Allows modifying object state, trigger events, or update components.

fn new(is_downed: bool) -> Self
where Self: Sized,

Construct a new instance of the script.

The is_downed parameter allows differentiating initial states or modes. Implementations can use this to customize setup logic.

fn clone_box(&self) -> Box<dyn Script + Send + Sync>