rusty_ache::engine::scene::game_object

Trait Object

pub trait Object {
    // Required methods
    fn new(
        components: Vec<Box<dyn Component + Send + Sync>>,
        script: Option<Box<dyn Script + Send + Sync>>,
        position: Position,
    ) -> Self;
    fn add_component(
        &mut self,
        component: Box<dyn Component + Send + Sync>,
    ) -> Result<(), GameObjectError>;
    fn remove_component(
        &mut self,
        component_id: usize,
    ) -> Result<(), GameObjectError>;
    fn get_position(&self) -> Result<&Position, GameObjectError>;
    fn update_position(
        &mut self,
        position: Position,
    ) -> Result<(), GameObjectError>;
    fn add_position(&mut self, vec: (i32, i32));
    fn run_action(&self);
}

Defines an interface for game objects.

Game objects are entities with components, position, and optional behavior scripts. This trait enumerates constructor and methods for component management, position queries and mutations, and script-driven actions.

Required Methods

fn new( components: Vec<Box<dyn Component + Send + Sync>>, script: Option<Box<dyn Script + Send + Sync>>, position: Position, ) -> Self

fn add_component( &mut self, component: Box<dyn Component + Send + Sync>, ) -> Result<(), GameObjectError>

fn remove_component( &mut self, component_id: usize, ) -> Result<(), GameObjectError>

fn get_position(&self) -> Result<&Position, GameObjectError>

fn update_position(&mut self, position: Position) -> Result<(), GameObjectError>

fn add_position(&mut self, vec: (i32, i32))

fn run_action(&self)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.