rusty_ache::engine

Trait Engine

pub trait Engine {
    // Required methods
    fn set_active_scene(&mut self, new_scene: Scene) -> Result<(), Error>;
    fn render(&mut self) -> Result<(), Error>;
    fn run(&mut self) -> Result<(), Error>;
    fn new(config: Box<dyn Config + Send>, scene: Scene) -> Self
       where Self: Sized;
}

Trait defining essential engine behavior.

Abstracts an engine capable of managing an active scene, performing rendering, running its main loop, and supporting dynamic configuration.

Required Methods

fn set_active_scene(&mut self, new_scene: Scene) -> Result<(), Error>

Sets the currently active scene within the engine.

fn render(&mut self) -> Result<(), Error>

Performs a rendering pass.

fn run(&mut self) -> Result<(), Error>

Starts and runs the engine main loop.

fn new(config: Box<dyn Config + Send>, scene: Scene) -> Self
where Self: Sized,

Creates a new engine instance from configuration and initial scene.